speed lookup_macro() up
Again, bloody awful code generation in even hotter spot - it's
called for every identifier.  What we do is scan the list of
symbols for given identifier, looking for the first one with
NS_MACRO or NS_UNDEF in sym->namespace, then check that it's
just NS_MACRO.  The thing is, NS_MACRO never mixes with anything
else, so instead of comparison we can just check that this bit
is set.  Result: test given bit in sym + constant offset vs.
fetch word, shift it, mask and compare.  Microoptimization,
but that thing is called often enough to give measurable effect.

Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>